tests: I0103 locate probe tail at probe_size offset
The probe response tail sits at offset probe_size in the writable
buffer, and probe_size is device specific and read from the device
config. The test hardcoded a 64 byte properties area and read the
tail at a fixed offset, so against a device advertising probe_size 24
it read the pre-filled canary instead of the real status and wrongly
saw 0xFF. Read probe_size from the config, size the writable
descriptor to probe_size plus the tail, and locate the tail at the
correct offset.

Also fix virtio_iommu_config.page_size_mask, which the spec defines
as le64 but was declared as a 32 bit field, misplacing probe_size in
the packed struct.

+++ lib/virtio_iommu.h
@@ -107,7 +107,7 @@ struct iommu_config {
 
 /* Full device configuration layout (spec 5.13.4). */
 struct virtio_iommu_config {
-    uint32_t page_size_mask;
+    uint64_t page_size_mask;
     struct {
         uint64_t start;
         uint64_t end;
